Added support for object detection (#10) object detection: UI and csv file updated added new headers for csv added command line arguments to select specific detection type fixed identity enabling when --identity is passed in new listener class for object
