+++ python-sdk-samples/display_metrics.py
@@ -2,17 +2,16 @@
 import cv2
 import os
 import math
-from collections import defaultdict
 
 TEXT_SIZE = 0.6
 PADDING_FOR_SEPARATOR = 5
 THRESHOLD_VALUE_FOR_EMOTIONS = 5
 
-identity_names_dict = defaultdict()
 
 IMAGES_DIR = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'images')
 
-def draw_metrics(args, frame, listener_metrics):
+def draw_metrics(frame, listener_metrics, identity_names_dict):
+
     """
     write metrics on screen
  
@@ -36,13 +35,14 @@ def draw_metrics(args, frame, listener_metrics):
         upper_right_x = upper_left_x + box_width
         upper_right_y = upper_left_y
 
-        if args.show_identity:
-            display_identity_on_screen(frame, listener_metrics["identities"][fid], upper_left_y, upper_left_x)
+        if "identities" in listener_metrics:
+            print(listener_metrics["identities"][fid])
+            display_identity_on_screen(frame, listener_metrics["identities"][fid], upper_left_y, upper_left_x, identity_names_dict)
  
         for key, val in measurements.items():
             display_measurements_on_screen(key, val, upper_left_y, frame, upper_left_x)
             upper_left_y += 25
- 
+
         for key, val in emotions.items():
             display_left_metrics(key.name, val, upper_left_x, upper_left_y, frame)
             upper_left_y += 25
@@ -61,16 +61,45 @@ def draw_metrics(args, frame, listener_metrics):
 
         display_left_metrics("glasses", listener_metrics["glasses"][fid], upper_left_x, upper_left_y, frame)
         upper_left_y += 25
- 
+
         for key, val in expressions.items():
             display_expressions_on_screen(key, val, upper_right_x, upper_right_y, frame)
             upper_right_y += 25
+        # draw gaze region for the first face. (ideally we should draw it for the driver only)
+        gaze_metrics = listener_metrics["gaze_metric"]
+        if len(gaze_metrics):
+            metric = next(iter(gaze_metrics.values()))
+            draw_gaze_region(frame, metric)
 
-    # draw gaze region for the first face. (ideally we should draw it for the driver only)
-    gaze_metrics = listener_metrics["gaze_metric"]
-    if len(gaze_metrics):
-        metric = next(iter(gaze_metrics.values()))
-        draw_gaze_region(frame, metric)
+            
+def draw_objects(frame, listener_metrics):
+    """
+    write objects with its bounding box on screen
+
+        Parameters
+        ----------
+        frame: affvisionpy.Frame
+            frame to write the metrics on
+
+        listener_metrics: dict
+            dictionary of dictionaries, gives current listener state
+
+    """
+    if "object_type" in listener_metrics:
+        for oid in listener_metrics["object_type"].keys():
+            upper_left_x, upper_left_y, lower_right_x, lower_right_y = get_bounding_box_points(oid, listener_metrics[
+                "bounding_box"])
+            box_width = lower_right_x - upper_left_x
+            upper_right_x = upper_left_x + box_width
+            upper_right_y = upper_left_y
+
+            cv2.rectangle(frame, (upper_left_x, upper_left_y), (lower_right_x, lower_right_y), (0, 0, 255), 3)
+
+            obj_type = listener_metrics["object_type"][oid].name
+            cv2.putText(frame, obj_type, (upper_right_x, upper_right_y), cv2.FONT_HERSHEY_DUPLEX, TEXT_SIZE,
+                        (0, 0, 0), 2, cv2.LINE_AA)
+            cv2.putText(frame, obj_type, (upper_right_x, upper_right_y), cv2.FONT_HERSHEY_DUPLEX, TEXT_SIZE,
+                        (255, 255, 255), 1, cv2.LINE_AA)
 
 
 def get_bounding_box_points(fid, bounding_box_dict):
@@ -91,8 +120,7 @@ def get_bounding_box_points(fid, bounding_box_dict):
             int(bounding_box_dict[fid][1]),
             int(bounding_box_dict[fid][2]),
             int(bounding_box_dict[fid][3]))
- 
- 
+
 def draw_bounding_box(frame, listener_metrics):
     """
     For each frame, draw the bounding box on screen.
@@ -122,7 +150,6 @@ def draw_bounding_box(frame, listener_metrics):
             cv2.rectangle(frame, (upper_left_x, upper_left_y), (lower_right_x, lower_right_y), (0, 255, 0), 3)
         else:
             cv2.rectangle(frame, (upper_left_x, upper_left_y), (lower_right_x, lower_right_y), (21, 169, 167), 3)
- 
 
 def get_text_size(text, font, thickness):
     """
@@ -144,8 +171,7 @@ def get_text_size(text, font, thickness):
     """
     text_size = cv2.getTextSize(text, font, TEXT_SIZE, thickness)
     return text_size[0][0], text_size[0][1]
- 
- 
+
 def display_measurements_on_screen(key, val, upper_left_y, frame, x1):
     """
     Display the measurement metrics on screen.
@@ -170,9 +196,9 @@ def display_measurements_on_screen(key, val, upper_left_y, frame, x1):
         val = 0
     val_text = str(round(val, 2))
     val_text_width, val_text_height = get_text_size(val_text, cv2.FONT_HERSHEY_SIMPLEX, 1)
- 
+
     key_val_width = key_text_width + val_text_width
- 
+
     cv2.putText(frame, key_name + ": ", (abs(x1 - key_val_width - PADDING_FOR_SEPARATOR), upper_left_y),
                 cv2.FONT_HERSHEY_SIMPLEX,
                 TEXT_SIZE,
@@ -180,7 +206,6 @@ def display_measurements_on_screen(key, val, upper_left_y, frame, x1):
     cv2.putText(frame, val_text, (abs(x1 - val_text_width), upper_left_y),
                 cv2.FONT_HERSHEY_SIMPLEX, TEXT_SIZE,
                 (255, 255, 255))
- 
 
 def display_left_metrics(key_name, val, upper_left_x, upper_left_y, frame):
     """
@@ -201,30 +226,30 @@ def display_left_metrics(key_name, val, upper_left_x, upper_left_y, frame):
  
     """
     key_text_width, key_text_height = get_text_size(key_name, cv2.FONT_HERSHEY_SIMPLEX, 1)
- 
+
     val_rect_width = 120
     key_val_width = key_text_width + val_rect_width
     cv2.putText(frame, key_name + ": ", (abs(upper_left_x - key_val_width), upper_left_y),
                 cv2.FONT_HERSHEY_SIMPLEX,
                 TEXT_SIZE,
-                (0, 0,0), 4, cv2.LINE_AA)
+                (0, 0, 0), 4, cv2.LINE_AA)
     cv2.putText(frame, key_name + ": ", (abs(upper_left_x - key_val_width), upper_left_y),
                 cv2.FONT_HERSHEY_SIMPLEX,
                 TEXT_SIZE,
                 (255, 255, 255), 2, cv2.LINE_AA)
     overlay = frame.copy()
- 
+
     if math.isnan(val):
         val = 0
- 
+
     start_box_point_x = abs(upper_left_x - val_rect_width)
     width = 8
     height = 10
- 
+
     rounded_val = roundup(val)
     rounded_val /= 10
     rounded_val = abs(int(rounded_val))
- 
+
     for i in range(0, rounded_val):
         start_box_point_x += 10
         cv2.rectangle(overlay, (start_box_point_x, upper_left_y),
@@ -239,11 +264,10 @@ def display_left_metrics(key_name, val, upper_left_x, upper_left_y, frame):
         start_box_point_x += 10
         cv2.rectangle(overlay, (start_box_point_x, upper_left_y),
                       (start_box_point_x + width, upper_left_y - height), (186, 186, 186), -1)
- 
+
     alpha = 0.8
     cv2.addWeighted(overlay, alpha, frame, 1 - alpha, 0, frame)
- 
- 
+
 def display_expressions_on_screen(key, val, upper_right_x, upper_right_y, frame):
     """
     Display the expressions metrics on screen.
@@ -267,16 +291,16 @@ def display_expressions_on_screen(key, val, upper_right_x, upper_right_y, frame)
     overlay = frame.copy()
     if math.isnan(val):
         val = 0
- 
+
     if 'blink' not in key_name:
         start_box_point_x = upper_right_x
         width = 8
         height = 10
- 
+
         rounded_val = roundup(val)
         rounded_val /= 10
         rounded_val = int(rounded_val)
-        for i in range(0, rounded_val ):
+        for i in range(0, rounded_val):
             start_box_point_x += 10
             cv2.rectangle(overlay, (start_box_point_x, upper_right_y),
                           (start_box_point_x + width, upper_right_y - height), (186, 186, 186), -1)
@@ -286,7 +310,7 @@ def display_expressions_on_screen(key, val, upper_right_x, upper_right_y, frame)
             start_box_point_x += 10
             cv2.rectangle(overlay, (start_box_point_x, upper_right_y),
                           (start_box_point_x + width, upper_right_y - height), (186, 186, 186), -1)
- 
+
         alpha = 0.8
         cv2.addWeighted(overlay, alpha, frame, 1 - alpha, 0, frame)
     else:
@@ -294,16 +318,74 @@ def display_expressions_on_screen(key, val, upper_right_x, upper_right_y, frame)
                     (0, 0, 0), 2, cv2.LINE_AA)
         cv2.putText(frame, str(val), (upper_right_x, upper_right_y), cv2.FONT_HERSHEY_DUPLEX, TEXT_SIZE,
                     (255, 255, 255), 1, cv2.LINE_AA)
- 
+
     cv2.putText(frame, " :" + str(key_name), (upper_right_x + val_rect_width, upper_right_y), cv2.FONT_HERSHEY_DUPLEX,
                 TEXT_SIZE,
                 (0, 0, 0), 4, cv2.LINE_AA)
     cv2.putText(frame, " :" + str(key_name), (upper_right_x + val_rect_width, upper_right_y), cv2.FONT_HERSHEY_DUPLEX,
                 TEXT_SIZE,
                 (255, 255, 255), 1, cv2.LINE_AA)
- 
 
-def display_identity_on_screen(frame, identity, upper_left_y, upper_left_x):
+def display_confidence_on_screen(key, val, upper_left_x, upper_left_y, frame):
+    """
+    Display the confidence metrics on screen.
+
+        Parameters
+        ----------
+        key: str
+            Name of the confidence.
+        val: float
+            Value of the emotion.
+        upper_left_x: int
+            the upper_left_x co-ordinate of the bounding box
+        upper_left_y: int
+            the upper_left_y co-ordinate of the bounding box
+        frame: affvisionpy.Frame
+            Frame object to write the confidence on
+
+    """
+    key_text_width, key_text_height = get_text_size(key, cv2.FONT_HERSHEY_SIMPLEX, 1)
+
+    val_rect_width = 120
+    key_val_width = key_text_width + val_rect_width
+    cv2.putText(frame, key + ": ", (abs(upper_left_x - key_val_width), upper_left_y),
+                cv2.FONT_HERSHEY_SIMPLEX,
+                TEXT_SIZE,
+                (0, 0, 0), 4, cv2.LINE_AA)
+    cv2.putText(frame, key + ": ", (abs(upper_left_x - key_val_width), upper_left_y),
+                cv2.FONT_HERSHEY_SIMPLEX,
+                TEXT_SIZE,
+                (255, 255, 255), 2, cv2.LINE_AA)
+    overlay = frame.copy()
+
+    if math.isnan(val):
+        val = 0
+
+    start_box_point_x = abs(upper_left_x - val_rect_width)
+    width = 8
+    height = 10
+
+    rounded_val = roundup(val)
+    rounded_val /= 10
+    rounded_val = abs(int(rounded_val))
+
+    for i in range(0, rounded_val):
+        start_box_point_x += 10
+        cv2.rectangle(overlay, (start_box_point_x, upper_left_y),
+                      (start_box_point_x + width, upper_left_y - height), (186, 186, 186), -1)
+
+        cv2.rectangle(overlay, (start_box_point_x, upper_left_y),
+                      (start_box_point_x + width, upper_left_y - height), (0, 204, 102), -1)
+
+    for i in range(rounded_val, 10):
+        start_box_point_x += 10
+        cv2.rectangle(overlay, (start_box_point_x, upper_left_y),
+                      (start_box_point_x + width, upper_left_y - height), (186, 186, 186), -1)
+
+    alpha = 0.8
+    cv2.addWeighted(overlay, alpha, frame, 1 - alpha, 0, frame)
+
+def display_identity_on_screen(frame, identity, upper_left_y, upper_left_x, identity_names_dict):
     """
         Display the face identity metrics on screen.
 
@@ -330,7 +412,6 @@ def display_identity_on_screen(frame, identity, upper_left_y, upper_left_x):
     cv2.putText(frame, id_name, (upper_left_x, upper_left_y - 10), cv2.FONT_HERSHEY_DUPLEX, TEXT_SIZE, (255, 255, 255),
                 1, cv2.LINE_AA)
 
-
 def draw_affectiva_logo(frame, width, height):
     """
     Place logo on the screen
@@ -344,11 +425,13 @@ def draw_affectiva_logo(frame, width, height):
         height: int
            height of the frame
     """
+
     logo = cv2.imread(os.path.join(IMAGES_DIR, 'Final logo - RGB Magenta.png'))
+
     logo_width = int(width / 3)
     logo_height = int(height / 10)
     logo = cv2.resize(logo, (logo_width, logo_height))
- 
+
     y1, y2 = 0, logo_height
     x1, x2 = width - logo_width, width
     # Remove the white background from the logo so that only the word "Affectiva" is visible on screen
@@ -357,7 +440,6 @@ def draw_affectiva_logo(frame, width, height):
         color = logo[0:logo_height, 0:logo_width, c] * (1.0 - alpha)
         beta = frame[y1:y2, x1:x2, c] * (alpha)
         frame[y1:y2, x1:x2, c] = color + beta
- 
 
 def draw_gaze_region(frame, gaze_metric):
     idx = int(gaze_metric.gaze_region)
@@ -407,7 +489,6 @@ def check_bounding_box_outside(width, height, bounding_box_dict):
         if upper_left_x < 0 or lower_right_x > width or upper_left_y < 0 or lower_right_y > height:
             return True
         return False
- 
 
 def roundup(num):
     """
